# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Comprehensive Semantic Analysis Web Service.",
        "title": "Content Creator Inshight",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "Returns pong , just to check server is running",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ping"
                ],
                "summary": "Get ping response",
                "responses": {
                    "200": {
                        "description": "pong",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/yt/analyse/video/{video_id}": {
            "get": {
                "description": "Returns Sentiment analyse of all video comments",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Yt"
                ],
                "summary": "list analyse detail of video comments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "videoID",
                        "name": "video_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "pong",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/yt/list/videos/{channelID}": {
            "get": {
                "description": "Returns list of  Youtube videos with channel id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Yt"
                ],
                "summary": "list youtube videos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "channelId",
                        "name": "channel_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "pong",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}